Started on new thing

+++ src/rom.h
@@ -1,3 +1,4 @@
+
 #ifndef __ROM_H__
 #define __ROM_H__
 
@@ -6,6 +7,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void rom_init(uint16_t base, uint16_t size, FILE *rom);
+void rom_init(uint32_t base, uint32_t size, FILE *rom);
 
 #endif // __ROM_H__
